chore: default value of representative_damage_percentile set to 100 to get the max damage to each Ci. Ci are low and high water flow speed for different road types. In total 6 Ci's for OSdamage

+++ ra2ce/analysis/direct/damage_calculation/damage_network_base.py
@@ -365,7 +365,8 @@ class DamageNetworkBase(ABC):
                     )
                 )
                 df[f"dam_{curve_name}_{event}_representative"] = (
-                    df.apply(lambda row: interpolate_damage(row, self.representative_damage_percentile), axis=1))
+                    df.apply(lambda row: interpolate_damage(row, self.representative_damage_percentile),
+                             axis=1)).astype(float)
 
                 # And throw way all intermediate results (that are not in the tuple)
                 df = df.drop(
